+++ packages/tracing/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/tresdoce/tresdoce-nestjs-toolkit/compare/@tresdoce-nestjs-toolkit/tracing@0.1.1...@tresdoce-nestjs-toolkit/tracing@0.1.2) (2023-07-17)
+
+**Note:** Version bump only for package @tresdoce-nestjs-toolkit/tracing
+
+## [0.1.1](https://github.com/tresdoce/tresdoce-nestjs-toolkit/compare/@tresdoce-nestjs-toolkit/tracing@0.1.0...@tresdoce-nestjs-toolkit/tracing@0.1.1) (2023-07-08)
+
+### 🐛 Bug Fixes
+
+- **tracing:** fixed versions of opentelemetry ([37c7d43](https://github.com/tresdoce/tresdoce-nestjs-toolkit/commit/37c7d43985754254cbbe63b515b2d1e0de422cf8))
+
+### 📝 Docs
+
+- edit readme ([5cf8db6](https://github.com/tresdoce/tresdoce-nestjs-toolkit/commit/5cf8db655ec587b6b52b8a9cc3cd283964ccd9bc))
+
 ## 0.1.0 (2023-06-13)
 
 ### 🧪 Tests